updated address request calls
