+++ index.ts
@@ -1,3 +1,4 @@
 export * from './src/address/index';
 export * from './src/provider';
 export * from './src/call';
+export * from './src/transactions/signPsbt';
